Check Exchange Logs Space
This is a script that can check the remaining disk space of any disk on any computer and then email a notification. I use this script together with Windows Task Scheduler to check the remaining space left on the logs drive of an Exchange server to keep it from filling up. When the logs drive fills up, Exchange services are no longer available. This script helps monitor that space to prevent an Exchange outage